ABS - Specify repo or package to download on command-line
This patch fixes the ability to specify which repos to be
synced as parameters to abs (e.g "abs core extra").

In addition, single or multiple packages can be requested
on the command line (e.g. "abs core/pacman extra/abs") as
well as any combination of repos and packages.

Signed-off-by: Allan McRae <allan@archlinux.org>